requirements.txt
streamlit
pandas
numpy
requests
joblib
xgboost
scikit-learn
matplotlib.pyplot as plt 
